# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  slug: string;
  title: string;
  location: string;
  size: string;
  category: string;
  zone: string;
  area: string;
  status: string;
  type: string;
  price: string;
  image: string;
  description: string;
}

export interface PropertyInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  size: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  zone: string;
  /** @minLength 1 */
  area: string;
  /** @minLength 1 */
  status: string;
  /** @minLength 1 */
  type: string;
  /** @minLength 1 */
  price: string;
  /** @minLength 1 */
  image: string;
  /** @minLength 1 */
  description: string;
}

export interface Inquiry {
  id: number;
  kind: string;
  name: string;
  contact: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  propertySlug?: string | null;
  message: string;
  createdAt: string;
}

export type InquiryInputKind = typeof InquiryInputKind[keyof typeof InquiryInputKind];


export const InquiryInputKind = {
  contact: 'contact',
  land: 'land',
} as const;

export interface InquiryInput {
  kind: InquiryInputKind;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  contact: string;
  location?: string;
  propertySlug?: string;
  /** @minLength 1 */
  message: string;
}

export type ListPropertiesParams = {
category?: string;
zone?: string;
status?: string;
type?: string;
};