# lib/api-zod/src/generated/types/inquiry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Inquiry {
  id: number;
  kind: string;
  name: string;
  contact: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  propertySlug?: string | null;
  message: string;
  createdAt: Date;
}