# lib/api-zod/src/generated/types/inquiryInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InquiryInputKind } from './inquiryInputKind';

export interface InquiryInput {
  kind: InquiryInputKind;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  contact: string;
  location?: string;
  propertySlug?: string;
  /** @minLength 1 */
  message: string;
}