# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  slug: string;
  title: string;
  location: string;
  size: string;
  category: string;
  zone: string;
  area: string;
  status: string;
  type: string;
  price: string;
  image: string;
  description: string;
}