# lib/api-zod/src/generated/types/propertyInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  location: string;
  /** @minLength 1 */
  size: string;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  zone: string;
  /** @minLength 1 */
  area: string;
  /** @minLength 1 */
  status: string;
  /** @minLength 1 */
  type: string;
  /** @minLength 1 */
  price: string;
  /** @minLength 1 */
  image: string;
  /** @minLength 1 */
  description: string;
}